+++ pkg/errformatter/service_valued_with_defaults_test_new_error_test.go
@@ -114,3 +114,165 @@ func TestServiceValuedWithDefaults_NewError(t *testing.T) {
 		}
 	})
 }
+
+func TestServiceValuedWithDefaults_NewError_ReWrap(t *testing.T) {
+	runName1 := "valued newError - error with 3 scope values for overwrite and details value for overwrite plus " +
+		"re-wrap by same fmt service"
+	t.Run(runName1, func(t *testing.T) {
+		var (
+			expectedResult = "valued_err_scope: some error text -> " +
+				"err_detail_true_1"
+
+			expectedReWrapedResult = "valued_err_scope: some error text -> " +
+				"err_detail_true_1, err_detail_true_2"
+		)
+
+		svc := NewValuesErrorFormatter([]Value{
+			{
+				num: KindScope,
+				any: "wrong_err_scope_for_overwrite",
+			},
+			{
+				num: KindScope,
+				any: "wrong_err_scope_2_for_overwrite",
+			},
+			{
+				num: KindScope,
+				any: "valued_err_scope",
+			},
+			{
+				num: KindDetails,
+				any: []string{"err_detail_true_1"},
+			},
+		}...)
+
+		err := svc.NewError("some error text")
+		if err.Error() != expectedResult {
+			t.Errorf("error text not equal with expected. current: %s, expected: %s",
+				err.Error(), expectedResult)
+		}
+
+		reWrapedErr := svc.ErrorOnly(err, "err_detail_true_2")
+		if reWrapedErr.Error() != expectedReWrapedResult {
+			t.Errorf("error text not equal with expected. current: %s, expected: %s",
+				reWrapedErr.Error(), expectedReWrapedResult)
+		}
+	})
+
+	runName2 := "valued error - error with one scope value, one details value and one code value with re-wrap by same service"
+	t.Run(runName2, func(t *testing.T) {
+		var (
+			expectedResult = "valued_err_scope: detail_info_as_value_1, detail_info_as_value_2 -> " +
+				"detail_in_value_1, detail_in_value_2"
+			expectedCode            = 9999
+			expectedReWrappedResult = "valued_err_scope: detail_info_as_value_1, detail_info_as_value_2 -> " +
+				"detail_in_value_1, detail_in_value_2"
+			reWrappedExpectedCode = 8765
+		)
+
+		svc := NewValuesErrorFormatter([]Value{
+			{
+				num: KindScope,
+				any: "valued_err_scope",
+			},
+			{
+				num: KindDetails,
+				any: []string{"detail_in_value_1", "detail_in_value_2"},
+			},
+			{
+				num: KindCode,
+				any: 100503,
+			},
+			{
+				num: KindCode,
+				any: expectedCode,
+			},
+		}...)
+
+		err := svc.NewError("detail_info_as_value_1", "detail_info_as_value_2")
+		if err.Error() != expectedResult {
+			t.Errorf("error text not equal with expected. current: %s, expected: %s",
+				err.Error(), expectedResult)
+		}
+
+		if code := svc.ErrorGetCode(err); code != expectedCode {
+			t.Errorf("error code not equal with expected. current: %d, expected: %d",
+				code, expectedCode)
+		}
+
+		if code := ValuedErrorGetCode(err); code != expectedCode {
+			t.Errorf("error code not equal with expected. current: %d, expected: %d",
+				code, expectedCode)
+		}
+
+		reWrappedErr := svc.ErrorWithCode(err, reWrappedExpectedCode)
+		if reWrappedErr.Error() != expectedReWrappedResult {
+			t.Errorf("error text not equal with expected. current: %s, expected: %s",
+				reWrappedErr.Error(), expectedReWrappedResult)
+		}
+
+		if code := svc.ErrorGetCode(err); code != reWrappedExpectedCode {
+			t.Errorf("error code not equal with expected. current: %d, expected: %d",
+				code, reWrappedExpectedCode)
+		}
+	})
+
+	runName3 := "valued error - error with one scope value, one details value and one code value with re-wrap by other service"
+	t.Run(runName3, func(t *testing.T) {
+		var (
+			expectedResult = "valued_err_scope: detail_info_as_value_1, detail_info_as_value_2 -> " +
+				"detail_in_value_1"
+			expectedCode            = 9999
+			expectedReWrappedResult = "re_wrap_scope: valued_err_scope: detail_info_as_value_1, detail_info_as_value_2 -> " +
+				"detail_in_value_1 -> re_wrap_detail_in_value_1"
+		)
+
+		svc := NewValuesErrorFormatter([]Value{
+			{
+				num: KindScope,
+				any: "valued_err_scope",
+			},
+			{
+				num: KindDetails,
+				any: []string{"detail_in_value_1"},
+			},
+			{
+				num: KindCode,
+				any: 100503,
+			},
+			{
+				num: KindCode,
+				any: expectedCode,
+			},
+		}...)
+
+		err := svc.NewError("detail_info_as_value_1", "detail_info_as_value_2")
+		if err.Error() != expectedResult {
+			t.Errorf("error text not equal with expected. current: %s, expected: %s",
+				err.Error(), expectedResult)
+		}
+
+		if code := svc.ErrorGetCode(err); code != expectedCode {
+			t.Errorf("error code not equal with expected. current: %d, expected: %d",
+				code, expectedCode)
+		}
+
+		if code := ValuedErrorGetCode(err); code != expectedCode {
+			t.Errorf("error code not equal with expected. current: %d, expected: %d",
+				code, expectedCode)
+		}
+
+		reWrapSvc := NewValuesErrorFormatter([]Value{
+			{
+				num: KindScope,
+				any: "re_wrap_scope",
+			},
+		}...)
+
+		reWrappedErr := reWrapSvc.Error(err, "re_wrap_detail_in_value_1")
+		if reWrappedErr.Error() != expectedReWrappedResult {
+			t.Errorf("error text not equal with expected. current: %s, expected: %s",
+				reWrappedErr.Error(), expectedReWrappedResult)
+		}
+	})
+}